% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jet.r
\name{fromEDN}
\alias{fromEDN}
\title{Function to convert an object from [EDN - Extensible Data Notation](https://github.com/edn-format/edn) format to R}
\usage{
fromEDN(edn_string)
}
\arguments{
\item{edn_string}{A character string in EDN format}
}
\value{
Returns an R object
}
\description{
Function to convert an object from [EDN - Extensible Data Notation](https://github.com/edn-format/edn) format to R
}
