\name{redn-package}
\alias{redn-package}
\alias{redn}
\docType{package}
\title{
\packageTitle{redn}
}
\description{
\packageDescription{redn}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{redn}
\packageIndices{redn}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{redn}

Maintainer: \packageMaintainer{redn}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
