% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jet.r
\name{toEDN}
\alias{toEDN}
\title{Function to convert an object to [EDN - Extensible Data Notation](https://github.com/edn-format/edn) format.}
\usage{
toEDN(r_obj)
}
\arguments{
\item{r_obj}{A R object to represent as EDN.}
}
\value{
Returns EDN string.
}
\description{
Function to convert an object to [EDN - Extensible Data Notation](https://github.com/edn-format/edn) format.
}
